Ultimo commit funciona todo

+++ database/migrations/2025_01_27_152532_create_events_table.php
@@ -16,10 +16,9 @@ return new class extends Migration
             $table->string('title');
             $table->foreignId('product_id')->constrained()->onDelete('cascade');
             $table->integer('units')->nullable();
-            $table->double('price',8,2)->nullable();
+            $table->decimal('price', 10, 2)->nullable();
             $table->dateTime('start');
             $table->dateTime('end')->nullable();
-            
             $table->timestamps();
         });
     }
